move curve data to separate file
