+++ src/asset-swapper/utils/market_operation_utils/bridge_source_utils.ts
@@ -2,25 +2,12 @@ import { ChainId } from '@0x/contract-addresses';
 import { BigNumber } from '@0x/utils';
 
 import {
-    ACRYPTOS_BSC_INFOS,
     APESWAP_ROUTER_BY_CHAIN_ID,
     BAKERYSWAP_ROUTER_BY_CHAIN_ID,
-    BELT_BSC_INFOS,
     BISWAP_ROUTER_BY_CHAIN_ID,
     COMPONENT_POOLS_BY_CHAIN_ID,
     CRYPTO_COM_ROUTER_BY_CHAIN_ID,
-    CURVE_V2_ARBITRUM_INFOS,
-    CURVE_AVALANCHE_INFOS,
-    CURVE_FANTOM_INFOS,
-    CURVE_MAINNET_INFOS,
-    CURVE_OPTIMISM_INFOS,
-    CURVE_POLYGON_INFOS,
-    CURVE_V2_AVALANCHE_INFOS,
-    CURVE_V2_FANTOM_INFOS,
-    CURVE_V2_MAINNET_INFOS,
-    CURVE_V2_POLYGON_INFOS,
     DFYN_ROUTER_BY_CHAIN_ID,
-    ELLIPSIS_BSC_INFOS,
     FIREBIRDONESWAP_BSC_INFOS,
     FIREBIRDONESWAP_POLYGON_INFOS,
     IRONSWAP_POLYGON_INFOS,
@@ -59,6 +46,21 @@ import {
     WAULTSWAP_ROUTER_BY_CHAIN_ID,
     YOSHI_ROUTER_BY_CHAIN_ID,
 } from './constants';
+import {
+    ACRYPTOS_BSC_INFOS,
+    BELT_BSC_INFOS,
+    CURVE_V2_ARBITRUM_INFOS,
+    CURVE_AVALANCHE_INFOS,
+    CURVE_FANTOM_INFOS,
+    CURVE_MAINNET_INFOS,
+    CURVE_OPTIMISM_INFOS,
+    CURVE_POLYGON_INFOS,
+    CURVE_V2_AVALANCHE_INFOS,
+    CURVE_V2_FANTOM_INFOS,
+    CURVE_V2_MAINNET_INFOS,
+    CURVE_V2_POLYGON_INFOS,
+    ELLIPSIS_BSC_INFOS,
+} from './curve';
 import { CurveInfo, PlatypusInfo } from './types';
 import { ERC20BridgeSource } from '../../types';
